changed capitalization of Styles to be styles.
